requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
gunicorn==21.2.0
Werkzeug==3.0.0